Change doctests and documentation
All the doctests now work correctly based on the target configuration
for the default compiler...I think. At least it all works on my
development machine (64-bit Windows 10).

More examples are added as doctests and the previous unit tests are
removed because these have become redundant to the new doctests.

Wording and typos are fixed for the comments as well.