Add command builder type
As more and more methods were being added with different combinations of
arguments in various locations, it became apparent that the Builder
pattern would be a better implementation and provide greater
flexibility. This hsa resulted in a radical change in the
implementation, documentation, and examples. However, it is in a much
better shape to address all of the problems that appeared when using the
previous implementation with another project.