Refactor to use nightly features
The `--print cfg` is now an optional argument for the `cargo rustc`
subcommand in the nightly toolchain. This whole crate will be refactored
to use that instead of the `cargo rustc -- --print cfg` hack. This
greatly simplifies the crate's implementation and reduces dependencies,
but it is a significant change and we need to watch for when the nightly
features that this crate uses becomes stable. The API is significantly
updated as well.