Add toolchain configuration
At the moment, this crate requires the nightly channel and the `-Z
unstable-options` feature to be enabled. Rather than hard-code these and
remove once the `--print` is stabilized, Cargo and rustup configuration
files are added to the project to indicate the requirement. Then, once
stabilization occurs, these files can be removed or updated without
changing the code.